[vector] Split SVG paint backend into hb-vector-paint-svg.cc
Extract all SVG-specific paint callbacks (push_transform,
push_clip_glyph, push_clip_rectangle, push_clip_path,
pop_clip, color, image, gradients, push/pop_group,
color_glyph, custom_palette_color) and the SVG render
function into a new file, matching the existing PDF split.

hb-vector-paint.cc retains the public API, format dispatch,
glyph_impl, clear, and render.

Also:
- Move ensure_initialized to a method on hb_vector_paint_t
- Move hb_vector_color_glyph_cache_key helper to .hh
- Rename hb_vector_paint_funcs_get → hb_vector_paint_svg_funcs_get
