+++ src/meson.build
@@ -731,6 +731,7 @@ hb_vector_sources = files(
   'hb-vector-draw.cc',
   'hb-vector-path.cc',
   'hb-vector-paint.cc',
+  'hb-vector-paint-svg.cc',
   'hb-vector-paint-pdf.cc',
   'hb-vector-path.hh',
   'hb-vector-svg-utils.cc',
